requirements.txt
streamlit
openai
requests
beautifulsoup4
pytrends
python-dotenv
